move dumpDatagram to tools
