+++ packages/multicast_dns/lib/multicast_dns.dart
@@ -185,22 +185,3 @@ class MDnsClient {
     }
   }
 }
-
-void _dumpDatagram(Datagram datagram) {
-  String _toHex(List<int> ints) {
-    final StringBuffer buffer = StringBuffer();
-    for (int i = 0; i < ints.length; i++) {
-      buffer.write(ints[i].toRadixString(16).padLeft(2, '0'));
-      if ((i + 1) % 10 == 0) {
-        buffer.writeln();
-      } else {
-        buffer.write(' ');
-      }
-    }
-    return buffer.toString();
-  }
-
-  print('${datagram.address.address}:${datagram.port}:');
-  print(_toHex(datagram.data));
-  print('');
-}
